PhongHNT update statistic

+++ src/main/java/vn/codegym/projections/IDayCostProjection.java
@@ -2,5 +2,5 @@ package vn.codegym.projections;
 
 public interface IDayCostProjection {
     String getDay();
-    Integer getCost();
+    Long getCost();
 }
